refactor: move Validation.java to testclasses
